backend/requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
numpy
python-jose